#39 - move interfaces to models/generics folder, update pg-promise to get access to IOptions
