+++ server/package.json
@@ -46,7 +46,6 @@
     "@types/cors": "~0.0.33",
     "@types/express": "~4.0.34",
     "@types/mocha": "~2.2.33",
-    "@types/pg-promise": "~0.0.28",
     "@types/sinon": "~1.16.33",
     "@types/bcrypt": "~1.0.0",
     "@types/jsonwebtoken": "~7.1.33",
@@ -60,7 +59,7 @@
     "chai-http": "~3.0.0",
     "mocha": "~3.2.0",
     "node-mocks-http": "~1.5.5",
-    "pg-promise": "~5.4.7",
+    "pg-promise": "~5.5.7",
     "sinon": "~1.17.6",
     "winston-spy": "~0.2.0",
     "bcrypt": "~1.0.1",
