+++ server/src/db/repositories/index.ts
@@ -1,0 +1,2 @@
+export * from "./recipe.repository";
+export * from "./user.repository";
