+++ server/src/models/generic/db-config.ts
@@ -1,0 +1,7 @@
+export interface DbConfig {
+  database: string;
+  host: string;
+  port: number;
+  user: string;
+  password: string;
+}
