+++ server/src/models/generic/db-extension.ts
@@ -1,0 +1,6 @@
+import { RecipeRepository, UserRepository } from "./../../db/repositories/index";
+
+export interface DbExtensions {
+    users: UserRepository;
+    recipes: RecipeRepository;
+}
