+++ server/src/models/index.ts
@@ -7,6 +7,8 @@ export * from "./errors/db-error";
 export * from "./errors/errors";
 export * from "./errors/error-type";
 export * from "./errors/validation-error";
+export * from "./generic/db-config";
+export * from "./generic/db-extension";
 export * from "./generic/socobo-request";
 export * from "./responses/login-response";
 export * from "./results/compare-pw-result";
